app/src/main/java/com/jameskbride/leastresistance/PathFinderFragment.java
package com.jameskbride.leastresistance;

import android.os.Bundle;
import android.support.annotation.Nullable;
import android.support.v4.app.Fragment;
import android.view.LayoutInflater;
import android.view.View;
import android.view.ViewGroup;

public class PathFinderFragment extends Fragment{
    public static PathFinderFragment newInstance() {
        return new PathFinderFragment();
    }

    @Nullable
    @Override
    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
        View view = inflater.inflate(R.layout.activity_least_resistance, container, false);

        return view;
    }
}
